+++ packages/components/forma-36-react-datepicker/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1-alpha.29](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.3.1-alpha.28...@contentful/forma-36-react-datepicker@0.3.1-alpha.29) (2020-08-28)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
 ## [0.3.1-alpha.28](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.3.1-alpha.27...@contentful/forma-36-react-datepicker@0.3.1-alpha.28) (2020-08-24)
 
 **Note:** Version bump only for package @contentful/forma-36-react-datepicker