+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.50.0](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.49.1...@contentful/forma-36-react-components@3.50.0) (2020-08-28)
+
+
+### Features
+
+* add modal launcher helper as alpha component ([#552](https://github.com/contentful/forma-36/issues/552)) ([a775798](https://github.com/contentful/forma-36/commit/a775798760ca5087e051b6093f73f1a42f1926a6))
+
+
+
+
+
 ## [3.49.1](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.49.0...@contentful/forma-36-react-components@3.49.1) (2020-08-24)
 
 **Note:** Version bump only for package @contentful/forma-36-react-components